Refactor daily question assignment workflow: remove prelims script and adjust main script for IST date handling
