+++ backend/assignDailyQuestion.js
@@ -26,7 +26,11 @@ if (!admin.apps.length) {
 const db = admin.firestore();
 
 async function assignDailyQuestion() {
-  const today = new Date().toISOString().substring(0, 10);
+  // Get IST date (UTC + 5:30)
+  const now = new Date();
+  const istOffset = 5.5 * 60 * 60 * 1000; // 5.5 hours in milliseconds
+  const istDate = new Date(now.getTime() + istOffset);
+  const today = istDate.toISOString().substring(0, 10);
 
   const todayDoc = await db
     .collection('daily_mains_questions')
@@ -64,4 +68,57 @@ async function assignDailyQuestion() {
   console.log('Assigned new daily question for', today, ':', qDoc.id);
 }
 
+async function assignDailyPrelimsQuestion() {
+  // Get IST date (UTC + 5:30)
+  const now = new Date();
+  const istOffset = 5.5 * 60 * 60 * 1000; // 5.5 hours in milliseconds
+  const istDate = new Date(now.getTime() + istOffset);
+  const today = istDate.toISOString().substring(0, 10);
+
+  // Check if today's daily question exists
+  const todayDoc = await db
+    .collection('daily_prelims_questions')
+    .doc(today)
+    .get();
+
+  if (todayDoc.exists) {
+    console.log('A Prelims question is already assigned for today.');
+    return;
+  }
+
+  // Gather all used questionIds
+  const usedDocs = await db.collection('daily_prelims_questions').get();
+  const usedIds = usedDocs.docs.map((doc) => doc.data().questionId);
+
+  // Get all possible questions
+  const allQs = (await db.collection('dataset/prelims/questions').get()).docs;
+
+  // Filter out used questions
+  const candidates = allQs.filter((q) => !usedIds.includes(q.id));
+  if (!candidates.length) {
+    throw new Error('No unused questions remaining!');
+  }
+
+  // Select a random question
+  const qDoc = candidates[Math.floor(Math.random() * candidates.length)];
+  const qData = qDoc.data();
+
+  // Create the day's doc in daily_prelims_questions
+  await db.collection('daily_prelims_questions').doc(today).set({
+    date: today,
+    questionId: qDoc.id,
+    Question: qData['Question and Year'],
+    Year: qData.Year,
+    Chapters: qData.Chapters,
+    Answer: qData.Answer,
+    Explanation: qData.Explanation,
+    Options: qData.Options,
+    Section: qData.Section,
+    Table: qData.Table,
+  });
+
+  console.log('Assigned new prelims daily question for', today, ':', qDoc.id);
+}
+
 assignDailyQuestion().catch(console.error);
+assignDailyPrelimsQuestion().catch(console.error);
